+++ javax/src/main/java/me/panpf/javax/util/Premisex.java
@@ -30,6 +30,9 @@ public class Premisex {
     }
 
 
+    /* ******************************************* boolean, null *******************************************/
+
+
     /**
      * Throws an [IllegalArgumentException] with [errorMessage] if the [value] is false.
      */
@@ -64,197 +67,436 @@ public class Premisex {
     }
 
 
+    /* ******************************************* file *******************************************/
+
+
     /**
-     * Throws a [FileNotFoundException] exception if the given file does not exist
+     * If the given file exist, it returns itself, otherwise it throws an FileNotFoundException
      */
-    public static void requireFileExist(@NotNull File file) throws FileNotFoundException {
-        if (!file.exists()) throw new FileNotFoundException(file.getPath());
+    public static File requireFileExist(@NotNull File file, @NotNull String paramName) throws FileNotFoundException {
+        if (file.exists()) {
+            return file;
+        } else {
+            throw new FileNotFoundException(String.format("The file pointed to by this parameter '%s': %s", paramName, file.getPath()));
+        }
     }
 
     /**
-     * Throws an [IllegalArgumentException] exception if the given file is not a directory
+     * If the given file exist, it returns itself, otherwise it throws an FileNotFoundException
      */
-    public static void requireIsDir(@NotNull File file) {
-        if (!file.isDirectory()) throw new IllegalArgumentException("Must be directory： " + file.getPath());
+    public static File requireFileExist(@NotNull File file) throws FileNotFoundException {
+        return requireFileExist(file, "unknown");
     }
 
     /**
-     * Throws an [IllegalArgumentException] exception if the given file is not a file
+     * If the given file is a directory, it returns itself, otherwise it throws an IllegalArgumentException
      */
-    public static void requireIsFile(@NotNull File file) {
-        if (!file.isFile()) throw new IllegalArgumentException("Must be file： " + file.getPath());
+    public static File requireIsDir(@NotNull File file, @NotNull String paramName) {
+        if (file.isDirectory()) {
+            return file;
+        } else {
+            throw new IllegalArgumentException(String.format("The file pointed to by the parameter '%s' is not a directory: %s", paramName, file.getPath()));
+        }
+    }
+
+    /**
+     * If the given file is a directory, it returns itself, otherwise it throws an IllegalArgumentException
+     */
+    public static File requireIsDir(@NotNull File file) {
+        return requireIsDir(file, "unknown");
+    }
+
+    /**
+     * If the given file is a file, it returns itself, otherwise it throws an IllegalArgumentException
+     */
+    public static File requireIsFile(@NotNull File file, @NotNull String paramName) {
+        if (file.isFile()) {
+            return file;
+        } else {
+            throw new IllegalArgumentException(String.format("The file pointed to by the parameter '%s' is not a file: %s", paramName, file.getPath()));
+        }
+    }
+
+    /**
+     * If the given file is a file, it returns itself, otherwise it throws an IllegalArgumentException
+     */
+    public static File requireIsFile(@NotNull File file) {
+        return requireIsFile(file, "unknown");
     }
 
 
+    /* ******************************************* range *******************************************/
+
+
     /**
-     * Returns true if [value] is within the range of [minValue] and [maxValue]
+     * If [value] is within the range of [minValue] and [maxValue], it returns itself, otherwise it throws an IllegalArgumentException
+     */
+    public static byte requireInRange(byte value, byte minValue, byte maxValue, @NotNull String paramName) {
+        if (value >= minValue && value <= maxValue) {
+            return value;
+        } else {
+            throw new IllegalArgumentException(String.format("The byte parameter '%s' value is %d, must be >= %d && <= %d", paramName, value, minValue, maxValue));
+        }
+    }
+
+    /**
+     * If [value] is within the range of [minValue] and [maxValue], it returns itself, otherwise it throws an IllegalArgumentException
      */
     public static byte requireInRange(byte value, byte minValue, byte maxValue) {
-        if (value < minValue || value > maxValue) {
-            throw new IllegalArgumentException(String.format("value %d must be >= %d && <= %d", value, minValue, maxValue));
-        }
-        return value;
+        return requireInRange(value, minValue, maxValue, "unknown");
     }
 
     /**
-     * Returns true if [value] is within the range of [minValue] and [maxValue]
+     * If [value] is within the range of [minValue] and [maxValue], it returns itself, otherwise it throws an IllegalArgumentException
+     */
+    public static short requireInRange(short value, short minValue, short maxValue, @NotNull String paramName) {
+        if (value >= minValue && value <= maxValue) {
+            return value;
+        } else {
+            throw new IllegalArgumentException(String.format("The short parameter '%s' value is %d, must be >= %d && <= %d", paramName, value, minValue, maxValue));
+        }
+    }
+
+    /**
+     * If [value] is within the range of [minValue] and [maxValue], it returns itself, otherwise it throws an IllegalArgumentException
      */
     public static short requireInRange(short value, short minValue, short maxValue) {
-        if (value < minValue || value > maxValue) {
-            throw new IllegalArgumentException(String.format("value %d must be >= %d && <= %d", value, minValue, maxValue));
-        }
-        return value;
+        return requireInRange(value, minValue, maxValue, "unknown");
     }
 
     /**
-     * Returns true if [value] is within the range of [minValue] and [maxValue]
+     * If [value] is within the range of [minValue] and [maxValue], it returns itself, otherwise it throws an IllegalArgumentException
+     */
+    public static int requireInRange(int value, int minValue, int maxValue, @NotNull String paramName) {
+        if (value >= minValue && value <= maxValue) {
+            return value;
+        } else {
+            throw new IllegalArgumentException(String.format("The int parameter '%s' value is %d, must be >= %d && <= %d", paramName, value, minValue, maxValue));
+        }
+    }
+
+    /**
+     * If [value] is within the range of [minValue] and [maxValue], it returns itself, otherwise it throws an IllegalArgumentException
      */
     public static int requireInRange(int value, int minValue, int maxValue) {
-        if (value < minValue || value > maxValue) {
-            throw new IllegalArgumentException(String.format("value %d must be >= %d && <= %d", value, minValue, maxValue));
-        }
-        return value;
+        return requireInRange(value, minValue, maxValue, "unknown");
     }
 
     /**
-     * Returns true if [value] is within the range of [minValue] and [maxValue]
+     * If [value] is within the range of [minValue] and [maxValue], it returns itself, otherwise it throws an IllegalArgumentException
+     */
+    public static long requireInRange(long value, long minValue, long maxValue, @NotNull String paramName) {
+        if (value >= minValue && value <= maxValue) {
+            return value;
+        } else {
+            throw new IllegalArgumentException(String.format("The long parameter '%s' value is %d, must be >= %d && <= %d", paramName, value, minValue, maxValue));
+        }
+    }
+
+    /**
+     * If [value] is within the range of [minValue] and [maxValue], it returns itself, otherwise it throws an IllegalArgumentException
      */
     public static long requireInRange(long value, long minValue, long maxValue) {
-        if (value < minValue || value > maxValue) {
-            throw new IllegalArgumentException(String.format("value %d must be >= %d && <= %d", value, minValue, maxValue));
-        }
-        return value;
+        return requireInRange(value, minValue, maxValue, "unknown");
     }
 
     /**
-     * Returns true if [value] is within the range of [minValue] and [maxValue]
+     * If [value] is within the range of [minValue] and [maxValue], it returns itself, otherwise it throws an IllegalArgumentException
+     */
+    public static float requireInRange(float value, float minValue, float maxValue, @NotNull String paramName) {
+        if (value >= minValue && value <= maxValue) {
+            return value;
+        } else {
+            throw new IllegalArgumentException(String.format("The float parameter '%s' value is %s, must be >= %s && <= %s", paramName, value, minValue, maxValue));
+        }
+    }
+
+    /**
+     * If [value] is within the range of [minValue] and [maxValue], it returns itself, otherwise it throws an IllegalArgumentException
      */
     public static float requireInRange(float value, float minValue, float maxValue) {
-        if (value < minValue || value > maxValue) {
-            throw new IllegalArgumentException(String.format("value %s must be >= %s && <= %s", value, minValue, maxValue));
-        }
-        return value;
+        return requireInRange(value, minValue, maxValue, "unknown");
     }
 
     /**
-     * Returns true if [value] is within the range of [minValue] and [maxValue]
+     * If [value] is within the range of [minValue] and [maxValue], it returns itself, otherwise it throws an IllegalArgumentException
+     */
+    public static double requireInRange(double value, double minValue, double maxValue, @NotNull String paramName) {
+        if (value >= minValue && value <= maxValue) {
+            return value;
+        } else {
+            throw new IllegalArgumentException(String.format("The double parameter '%s' value is %s, must be >= %s && <= %s", paramName, value, minValue, maxValue));
+        }
+    }
+
+    /**
+     * If [value] is within the range of [minValue] and [maxValue], it returns itself, otherwise it throws an IllegalArgumentException
      */
     public static double requireInRange(double value, double minValue, double maxValue) {
-        if (value < minValue || value > maxValue) {
-            throw new IllegalArgumentException(String.format("value %s must be >= %s && <= %s", value, minValue, maxValue));
-        }
-        return value;
+        return requireInRange(value, minValue, maxValue, "unknown");
     }
 
 
     /**
-     * Return true if [value] is not in the range [minValue] and [maxValue]
+     * If [value] is not in the range [minValue] and [maxValue], it returns itself, otherwise it throws an IllegalArgumentException
+     */
+    public static byte requireNotInRange(byte value, byte minValue, byte maxValue, @NotNull String paramName) {
+        if (value < minValue || value > maxValue) {
+            return value;
+        } else {
+            throw new IllegalArgumentException(String.format("The byte parameter '%s' value is %d, must be < %d || > %d", paramName, value, minValue, maxValue));
+        }
+    }
+
+    /**
+     * If [value] is not in the range [minValue] and [maxValue], it returns itself, otherwise it throws an IllegalArgumentException
      */
     public static byte requireNotInRange(byte value, byte minValue, byte maxValue) {
-        if (value >= minValue && value <= maxValue) {
-            throw new IllegalArgumentException(String.format("value %d must be < %d || > %d", value, minValue, maxValue));
-        }
-        return value;
+        return requireNotInRange(value, minValue, maxValue, "unknown");
     }
 
     /**
-     * Return true if [value] is not in the range [minValue] and [maxValue]
+     * If [value] is not in the range [minValue] and [maxValue], it returns itself, otherwise it throws an IllegalArgumentException
+     */
+    public static short requireNotInRange(short value, short minValue, short maxValue, @NotNull String paramName) {
+        if (value < minValue || value > maxValue) {
+            return value;
+        } else {
+            throw new IllegalArgumentException(String.format("The short parameter '%s' value is %d, must be < %d || > %d", paramName, value, minValue, maxValue));
+        }
+    }
+
+    /**
+     * If [value] is not in the range [minValue] and [maxValue], it returns itself, otherwise it throws an IllegalArgumentException
      */
     public static short requireNotInRange(short value, short minValue, short maxValue) {
-        if (value >= minValue && value <= maxValue) {
-            throw new IllegalArgumentException(String.format("value %d must be < %d || > %d", value, minValue, maxValue));
-        }
-        return value;
+        return requireNotInRange(value, minValue, maxValue, "unknown");
     }
 
     /**
-     * Return true if [value] is not in the range [minValue] and [maxValue]
+     * If [value] is not in the range [minValue] and [maxValue], it returns itself, otherwise it throws an IllegalArgumentException
+     */
+    public static int requireNotInRange(int value, int minValue, int maxValue, @NotNull String paramName) {
+        if (value < minValue || value > maxValue) {
+            return value;
+        } else {
+            throw new IllegalArgumentException(String.format("The int parameter '%s' value is %d, must be < %d || > %d", paramName, value, minValue, maxValue));
+        }
+    }
+
+    /**
+     * If [value] is not in the range [minValue] and [maxValue], it returns itself, otherwise it throws an IllegalArgumentException
      */
     public static int requireNotInRange(int value, int minValue, int maxValue) {
-        if (value >= minValue && value <= maxValue) {
-            throw new IllegalArgumentException(String.format("value %d must be < %d || > %d", value, minValue, maxValue));
-        }
-        return value;
+        return requireNotInRange(value, minValue, maxValue, "unknown");
     }
 
     /**
-     * Return true if [value] is not in the range [minValue] and [maxValue]
+     * If [value] is not in the range [minValue] and [maxValue], it returns itself, otherwise it throws an IllegalArgumentException
+     */
+    public static long requireNotInRange(long value, long minValue, long maxValue, @NotNull String paramName) {
+        if (value < minValue || value > maxValue) {
+            return value;
+        } else {
+            throw new IllegalArgumentException(String.format("The long parameter '%s' value is %d, must be < %d || > %d", paramName, value, minValue, maxValue));
+        }
+    }
+
+    /**
+     * If [value] is not in the range [minValue] and [maxValue], it returns itself, otherwise it throws an IllegalArgumentException
      */
     public static long requireNotInRange(long value, long minValue, long maxValue) {
-        if (value >= minValue && value <= maxValue) {
-            throw new IllegalArgumentException(String.format("value %d must be < %d || > %d", value, minValue, maxValue));
-        }
-        return value;
+        return requireNotInRange(value, minValue, maxValue, "unknown");
     }
 
     /**
-     * Return true if [value] is not in the range [minValue] and [maxValue]
+     * If [value] is not in the range [minValue] and [maxValue], it returns itself, otherwise it throws an IllegalArgumentException
+     */
+    public static float requireNotInRange(float value, float minValue, float maxValue, @NotNull String paramName) {
+        if (value < minValue || value > maxValue) {
+            return value;
+        } else {
+            throw new IllegalArgumentException(String.format("The float parameter '%s' value is %s, must be < %s || > %s", paramName, value, minValue, maxValue));
+        }
+    }
+
+    /**
+     * If [value] is not in the range [minValue] and [maxValue], it returns itself, otherwise it throws an IllegalArgumentException
      */
     public static float requireNotInRange(float value, float minValue, float maxValue) {
-        if (value >= minValue && value <= maxValue) {
-            throw new IllegalArgumentException(String.format("value %s must be < %s || > %s", value, minValue, maxValue));
-        }
-        return value;
+        return requireNotInRange(value, minValue, maxValue, "unknown");
     }
 
     /**
-     * Return true if [value] is not in the range [minValue] and [maxValue]
+     * If [value] is not in the range [minValue] and [maxValue], it returns itself, otherwise it throws an IllegalArgumentException
+     */
+    public static double requireNotInRange(double value, double minValue, double maxValue, @NotNull String paramName) {
+        if (value < minValue || value > maxValue) {
+            return value;
+        } else {
+            throw new IllegalArgumentException(String.format("The double parameter '%s' value is %s, must be < %s || > %s", paramName, value, minValue, maxValue));
+        }
+    }
+
+    /**
+     * If [value] is not in the range [minValue] and [maxValue], it returns itself, otherwise it throws an IllegalArgumentException
      */
     public static double requireNotInRange(double value, double minValue, double maxValue) {
-        if (value >= minValue && value <= maxValue) {
-            throw new IllegalArgumentException(String.format("value %s must be < %s || > %s", value, minValue, maxValue));
+        return requireNotInRange(value, minValue, maxValue, "unknown");
+    }
+
+
+    /* ******************************************* string safe *******************************************/
+
+
+    /**
+     * If the [value] is null or empty or blank, it returns itself, otherwise it throws an IllegalArgumentException.
+     */
+    public static <T extends CharSequence> T requireSafe(@Nullable T value, @NotNull String paramName) {
+        if (Stringx.isSafe(value)) {
+            return value;
+        } else {
+            throw new IllegalArgumentException(String.format("The string parameter '%s' is null or empty or blank", paramName));
         }
-        return value;
-    }
-
-
-    /**
-     * Throws an [IllegalArgumentException] with the result of calling [lazyMessage] if the [value] is not safe. Otherwise returns the value.
-     */
-    public static <T extends CharSequence> T requireSafe(@Nullable T value, @NotNull LazyValue<String> lazyMessage) {
-        if (!Stringx.isSafe(value)) throw new IllegalArgumentException(lazyMessage.get());
-        return value;
     }
 
     /**
-     * Throws an [IllegalArgumentException] with [errorMessage] if the [value] is not safe. Otherwise returns the value.
-     */
-    public static <T extends CharSequence> T requireSafe(@Nullable T value, @NotNull String errorMessage) {
-        if (!Stringx.isSafe(value)) throw new IllegalArgumentException(errorMessage);
-        return value;
-    }
-
-    /**
-     * Throws an [IllegalArgumentException] if the [value] is not safe. Otherwise returns the value.
+     * If the [value] is null or empty or blank, it returns itself, otherwise it throws an IllegalArgumentException.
      */
     public static <T extends CharSequence> T requireSafe(@Nullable T value) {
-        requireSafe(value, "Failed requireSafe.");
-        return value;
+        return requireSafe(value, "unknown");
     }
 
     /**
-     * Throws an [IllegalArgumentException] with the result of calling [lazyMessage] if the [value] is safe. Otherwise returns the value.
+     * If the [value] is not null or empty or blank, it returns itself, otherwise it throws an IllegalArgumentException.
      */
-    public static <T extends CharSequence> T requireNotSafe(@Nullable T value, @NotNull LazyValue<String> lazyMessage) {
-        if (Stringx.isSafe(value)) throw new IllegalArgumentException(lazyMessage.get());
-        return value;
+    public static <T extends CharSequence> T requireNotSafe(@Nullable T value, @NotNull String paramName) {
+        if (Stringx.isNotSafe(value)) {
+            return value;
+        } else {
+            throw new IllegalArgumentException(String.format("The string parameter '%s' is not null or empty or blank", paramName));
+        }
     }
 
     /**
-     * Throws an [IllegalArgumentException] with [errorMessage] if the [value] is safe. Otherwise returns the value.
-     */
-    public static <T extends CharSequence> T requireNotSafe(@Nullable T value, @NotNull String errorMessage) {
-        if (Stringx.isSafe(value)) throw new IllegalArgumentException(errorMessage);
-        return value;
-    }
-
-    /**
-     * Throws an [IllegalArgumentException] if the [value] is safe. Otherwise returns the value.
+     * If the [value] is not null or empty or blank, it returns itself, otherwise it throws an IllegalArgumentException.
      */
     public static <T extends CharSequence> T requireNotSafe(@Nullable T value) {
-        requireNotSafe(value, "Failed requireNotSafe.");
-        return value;
+        return requireNotSafe(value, "unknown");
     }
+
+
+    /* ******************************************* number zero *******************************************/
+
+
+    /**
+     * If [value] is not 0, it returns itself, otherwise it throws an exception.
+     */
+    public static byte requireNotZero(byte value, @NotNull String paramName) {
+        if (value != (byte) 0) {
+            return value;
+        } else {
+            throw new IllegalArgumentException(String.format("The parameter '%s' cannot be 0", paramName));
+        }
+    }
+
+    /**
+     * If [value] is not 0, it returns itself, otherwise it throws an exception.
+     */
+    public static byte requireNotZero(byte value) {
+        return requireNotZero(value, "unknown");
+    }
+
+    /**
+     * If [value] is not 0, it returns itself, otherwise it throws an exception.
+     */
+    public static short requireNotZero(short value, @NotNull String paramName) {
+        if (value != (short) 0) {
+            return value;
+        } else {
+            throw new IllegalArgumentException(String.format("The parameter '%s' cannot be 0", paramName));
+        }
+    }
+
+    /**
+     * If [value] is not 0, it returns itself, otherwise it throws an exception.
+     */
+    public static short requireNotZero(short value) {
+        return requireNotZero(value, "unknown");
+    }
+
+    /**
+     * If [value] is not 0, it returns itself, otherwise it throws an exception.
+     */
+    public static int requireNotZero(int value, @NotNull String paramName) {
+        if (value != 0) {
+            return value;
+        } else {
+            throw new IllegalArgumentException(String.format("The parameter '%s' cannot be 0", paramName));
+        }
+    }
+
+    /**
+     * If [value] is not 0, it returns itself, otherwise it throws an exception.
+     */
+    public static int requireNotZero(int value) {
+        return requireNotZero(value, "unknown");
+    }
+
+    /**
+     * If [value] is not 0, it returns itself, otherwise it throws an exception.
+     */
+    public static long requireNotZero(long value, @NotNull String paramName) {
+        if (value != 0L) {
+            return value;
+        } else {
+            throw new IllegalArgumentException(String.format("The parameter '%s' cannot be 0", paramName));
+        }
+    }
+
+    /**
+     * If [value] is not 0, it returns itself, otherwise it throws an exception.
+     */
+    public static long requireNotZero(long value) {
+        return requireNotZero(value, "unknown");
+    }
+
+    /**
+     * If [value] is not 0, it returns itself, otherwise it throws an exception.
+     */
+    public static float requireNotZero(float value, @NotNull String paramName) {
+        if (value != 0f) {
+            return value;
+        } else {
+            throw new IllegalArgumentException(String.format("The parameter '%s' cannot be 0", paramName));
+        }
+    }
+
+    /**
+     * If [value] is not 0, it returns itself, otherwise it throws an exception.
+     */
+    public static float requireNotZero(float value) {
+        return requireNotZero(value, "unknown");
+    }
+
+    /**
+     * If [value] is not 0, it returns itself, otherwise it throws an exception.
+     */
+    public static double requireNotZero(double value, @NotNull String paramName) {
+        if (value != 0.0) {
+            return value;
+        } else {
+            throw new IllegalArgumentException(String.format("The parameter '%s' cannot be 0", paramName));
+        }
+    }
+
+    /**
+     * If [value] is not 0, it returns itself, otherwise it throws an exception.
+     */
+    public static double requireNotZero(double value) {
+        return requireNotZero(value, "unknown");
+    }
+
 
 
     /*